Remap classes into better packaging.
