+++ src/main/java/io/github/theodoremeyer/spigotmc/simplevoicegeyser/server/servlets/AudioWorkletServlet.java
@@ -1,4 +1,4 @@
-package io.github.theodoremeyer.spigotmc.simplevoicegeyser.server;
+package io.github.theodoremeyer.spigotmc.simplevoicegeyser.server.servlets;
 
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
@@ -6,15 +6,14 @@ import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 /**
- * Class that makes the websocket client's mic work
- * Returns the js file.
+ * The Audio js file to play audio on the client.
  */
-public class MicWorkletServlet extends HttpServlet {
+public class AudioWorkletServlet extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException {
         resp.setContentType("application/javascript");
-        try (var in = getClass().getResourceAsStream("/web/mic-capture-processor.js")) {
+        try (var in = getClass().getResourceAsStream("/web/js/audio-worklet-processor.js")) {
             if (in == null) {
                 resp.setStatus(HttpServletResponse.SC_NOT_FOUND);
                 return;
