+++ src/main/java/io/github/theodoremeyer/spigotmc/simplevoicegeyser/server/servlets/JettyHtmlServlet.java
@@ -1,4 +1,4 @@
-package io.github.theodoremeyer.spigotmc.simplevoicegeyser.server;
+package io.github.theodoremeyer.spigotmc.simplevoicegeyser.server.servlets;
 
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
@@ -6,14 +6,20 @@ import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 /**
- * The Audio js file to play audio on the client.
+ * Almost Complete now!
+ * HTML server page
  */
-public class AudioWorkletServlet extends HttpServlet {
-
+public class JettyHtmlServlet extends HttpServlet {
+    /**
+     * HTML Page
+     * @param req request
+     * @param resp http-response
+     * @throws IOException Exception
+     */
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException {
-        resp.setContentType("application/javascript");
-        try (var in = getClass().getResourceAsStream("/web/audio-worklet-processor.js")) {
+        resp.setContentType("text/html");
+        try (var in = getClass().getResourceAsStream("/web/index.html")) {
             if (in == null) {
                 resp.setStatus(HttpServletResponse.SC_NOT_FOUND);
                 return;
