+++ src/main/java/io/github/theodoremeyer/spigotmc/simplevoicegeyser/server/servlets/MicWorkletServlet.java
@@ -1,4 +1,4 @@
-package io.github.theodoremeyer.spigotmc.simplevoicegeyser.server;
+package io.github.theodoremeyer.spigotmc.simplevoicegeyser.server.servlets;
 
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
@@ -6,20 +6,15 @@ import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 /**
- * Almost Complete now!
- * HTML server page
+ * Class that makes the websocket client's mic work
+ * Returns the js file.
  */
-public class JettyHtmlServlet extends HttpServlet {
-    /**
-     * HTML Page
-     * @param req request
-     * @param resp http-response
-     * @throws IOException Exception
-     */
+public class MicWorkletServlet extends HttpServlet {
+
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException {
-        resp.setContentType("text/html");
-        try (var in = getClass().getResourceAsStream("/web/index.html")) {
+        resp.setContentType("application/javascript");
+        try (var in = getClass().getResourceAsStream("/web/js/mic-capture-processor.js")) {
             if (in == null) {
                 resp.setStatus(HttpServletResponse.SC_NOT_FOUND);
                 return;
