+++ src/main/resources/web/js/mic-capture-processor.js
